+++ bin/cli/src/transact/blob.rs
@@ -1,4 +1,4 @@
-// Copyright 2024 RISC Zero, Inc.
+// Copyright 2024, 2025 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,16 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{retry, retry_with_context};
-use alloy::consensus::{Blob, BlobTransactionSidecar, BlockHeader};
+use crate::retry;
+use alloy::consensus::{Blob, BlobTransactionSidecar};
 use alloy::eips::eip4844::{kzg_to_versioned_hash, BLS_MODULUS, FIELD_ELEMENTS_PER_BLOB};
-use alloy::eips::{BlockId, BlockNumberOrTag};
-use alloy::network::{BlockResponse, Network};
-use alloy::primitives::{BlockNumber, B256, U256};
-use alloy::providers::Provider;
+use alloy::primitives::{B256, U256};
 use alloy_rpc_types_beacon::sidecar::{BeaconBlobBundle, BlobData};
-use anyhow::{anyhow, bail, Context};
-use kailua_client::{await_tel, await_tel_res};
+use anyhow::{bail, Context};
 use opentelemetry::global::tracer;
 use opentelemetry::trace::{FutureExt, TraceContextExt, Tracer};
 use serde::de::DeserializeOwned;
@@ -193,75 +189,4 @@ pub fn blob_fe_proof(
     } else {
         bail!("Generated invalid kzg proof.")
     }
-}
-
-pub async fn get_next_block<P: Provider<N>, N: Network>(
-    provider: P,
-    parent_hash: B256,
-) -> anyhow::Result<N::BlockResponse> {
-    let tracer = tracer("kailua");
-    let context = opentelemetry::Context::current_with_span(tracer.start("get_next_block"));
-
-    let block_parent = await_tel_res!(
-        context,
-        tracer,
-        "Provider::get_block_by_hash",
-        retry_with_context!(async {
-            provider
-                .get_block_by_hash(parent_hash)
-                .await
-                .context("get_block_by_hash")?
-                .ok_or_else(|| anyhow!("Failed to fetch parent block"))
-        })
-    )?;
-    let parent_number = block_parent.header().number();
-    let block = await_tel!(context, get_block_by_number(&provider, parent_number + 1))?;
-
-    Ok(block)
-}
-
-pub async fn get_block_by_number<P: Provider<N>, N: Network>(
-    provider: P,
-    block_number: BlockNumber,
-) -> anyhow::Result<N::BlockResponse> {
-    let tracer = tracer("kailua");
-    let context = opentelemetry::Context::current_with_span(tracer.start("get_block_by_number"));
-
-    let block = await_tel_res!(
-        context,
-        tracer,
-        "Provider::get_block_by_number",
-        retry_with_context!(async {
-            provider
-                .get_block_by_number(BlockNumberOrTag::Number(block_number))
-                .await
-                .context("get_block_by_number")?
-                .ok_or_else(|| anyhow!("Failed to fetch block"))
-        })
-    )?;
-
-    Ok(block)
-}
-
-pub async fn get_block<P: Provider<N>, N: Network>(
-    provider: P,
-    block_id: BlockNumberOrTag,
-) -> anyhow::Result<N::BlockResponse> {
-    let tracer = tracer("kailua");
-    let context = opentelemetry::Context::current_with_span(tracer.start("get_block"));
-
-    let block = await_tel_res!(
-        context,
-        tracer,
-        "Provider::get_block",
-        retry_with_context!(async {
-            provider
-                .get_block(BlockId::Number(block_id))
-                .await
-                .context("get_block")?
-                .ok_or_else(|| anyhow!("Failed to fetch block"))
-        })
-    )?;
-
-    Ok(block)
 }
